pyproject.toml
[tool.poetry]
name = "krautsalat-xyz"
version = "0.1.0"
description = ""
authors = ["Kraut Salat <kaputkrauts@outlook.com>"]
readme = "README.md"
packages = [{include = "krautsalat"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
pydantic = "^1.10.5"
motor = "^3.1.1"
beanie = "^1.17.0"
hypercorn = {extras = ["uvloop"], version = "^0.14.3"}
redis = "^4.5.1"
quart = "^0.18.3"
quart-schema = "^0.15.0"
quart-session = "^2.0.0"
aiocache = "^0.12.0"
email-validator = "^1.3.1"
python-slugify = "^8.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
